fix(cli-common): alkane amounts are u128 — u64 was freezing real balances
23.0137 LP is unspendable on mainnet right now. An 18-decimal token's sub-units
exceed u64::MAX above 18.4467 units, so every amount path that narrowed to u64
either refused to parse or wrapped to a silent zero, and the wallet reported
"Insufficient alkanes: need X, have 0" for coins it could plainly see.

Everything is u128 on chain. PrefetchedAlkane already documents `amount` as a
String precisely so it can round-trip u128 — the consumers narrowed it back.

Widened, with the reason each one mattered:

  types.rs        InputRequirement::Alkanes.amount and ProtostoneEdict.amount.
                  The edict one is not cosmetic: protorune's on-chain
                  Edict.amount is u128 and the runestone encoder varint-encodes
                  it, so a u64 edict truncated the transfer itself.
  parsing.rs      the three alkane amount parsers. Bitcoin amounts stay u64 —
                  sats fit, and widening them is churn with no payoff.
  execute.rs      the needed/found/collected/totals maps and the two JSON amount
                  readers, which now use the json_u128 helper that already
                  existed, unused, next to the narrow json_u64.
  amm_cli.rs      `params.amount0 as u64` on add-liquidity and swap. These were
                  the worst of them: params.amountN is ALREADY u128, so `as`
                  wrapped silently on the exact LP path that froze.
  web-sys         `e.amount as u64` in the edict mapping — the layer the app
                  actually calls.
  alkanes-cli     the swap input requirement.

The espo fallback also re-emitted the amount as a JSON *number*; serde_json
Numbers cannot carry u128, so it now emits a string, which is the form the
reader above it already accepted. Its `unwrap_or(0)` is literally what turned a
real balance into "have 0".

Map KEYS stay (u64, u64) — block/tx are ids and are not the bug. Only amounts
moved.

6 regression tests pin the exact frozen figure: 23.0137 at 18 decimals is
23_013_700_000_000_000_000, above u64::MAX (18_446_744_073_709_551_615). One
test asserts that premise so the others cannot go vacuous; others cover
u128::MAX, edicts in both spec forms, Bitcoin staying u64, and an overflowing
amount being a clean parse ERROR rather than another silent zero.

277 lib tests pass. The 25 errors in the tests/ integration targets are
pre-existing (stale struct fields, moved modules, a changed arity) — verified by
stashing this change and re-measuring: 25 before, 25 after.

STILL REQUIRED for the frozen LP to move: rebuild the WASM and sync
lib/oyl/alkanes/ in subfrost-app. This commit alone changes nothing for users.

+++ crates/alkanes-cli/src/main.rs
@@ -3604,7 +3604,9 @@ async fn execute_alkanes_command<T: System>(system: &mut T, command: Alkanes, fr
                 vec![InputRequirement::Alkanes {
                     block: input_token.block,
                     tx: input_token.tx,
-                    amount: input as u64,
+                    // No `as u64`: alkane amounts are u128. The Bitcoin arm above
+                    // keeps its cast because sats genuinely fit in u64.
+                    amount: input,
                 }]
             };
